tagged_verse_export / move methods to exportController
