+++ app/frontend/controllers/export_controller.js
@@ -16,6 +16,10 @@
    along with Ezra Bible App. See the file LICENSE.
    If not, see <http://www.gnu.org/licenses/>. */
 
+const i18nHelper = require('../helpers/i18n_helper.js');
+const marked = require('marked');
+
+
 /**
  * The ExportController implements the export of certain verses with notes or tags into a Word document.
  *
@@ -24,7 +28,6 @@
 
 
 var exportFilePath = null;
-
 
 module.exports.showSaveDialog = async function (fileTitle) {
   const dialog = require('electron').remote.dialog;
@@ -40,6 +43,71 @@ module.exports.showSaveDialog = async function (fileTitle) {
     }
   });
 };
+
+module.exports.saveWordDocument = async function(title, bibleBooks, verses) {
+  const officegen = require('officegen');
+  const fs = require('fs');
+  const shell = require('electron').remote.shell;
+
+  const titleWrapper = document.createElement('p');
+  titleWrapper.innerHTML = marked(title);
+  const titleParts = titleWrapper.firstElementChild.childNodes;
+
+  var docx = officegen({
+    type: 'docx',
+    title: titleWrapper.textContent,
+    description: 'Automatically generated by Ezra Bible App',
+    pageMargins: {
+      top: 1200,
+      bottom: 1200,
+      left: 1000,
+      right: 1000
+    }
+  });
+
+  // Officegen calling this function after finishing to generate the docx document:
+  docx.on('finalize', (written) => {
+    shell.openPath(exportFilePath);
+  });
+
+  // Officegen calling this function to report errors:
+  docx.on('error', function(err) {
+    console.log(err);
+  });
+  
+  var p = docx.createP();
+  
+  for(let i = 0; i < titleParts.length; i++) {
+    p.addText(titleParts[i].textContent, { font_size: 14, bold: true, italic: titleParts[i].nodeName == 'EM' });
+  }
+  p.addLineBreak();
+  p.addLineBreak();
+
+  for (let i = 0; i < bibleBooks.length; i++) {
+    const currentBook = bibleBooks[i];
+    const bookTitle = await i18nHelper.getSwordTranslation(currentBook.longTitle);
+
+    p.addText(bookTitle, { bold: true });
+    p.addLineBreak();
+
+    const allBlocks = getBibleBookVerseBlocks(currentBook, verses);
+    await renderVerseBlocks(p, currentBook, allBlocks);
+
+    // Line break after book end
+    p.addLineBreak();
+  }
+
+  //console.log("Generating word document " + this.saveFilePath);
+  var out = fs.createWriteStream(exportFilePath);
+
+  out.on('error', function(err) {
+    console.log(err);
+  });
+
+  // Async call to generate the output file:
+  docx.generate(out);
+};
+
 
 function getExportDialogOptions(title) {
   const app = require('electron').remote.app;
@@ -64,4 +132,85 @@ function getPaddedNumber(number) {
     paddedNumber = "0" + number;
   }
   return paddedNumber;
+}
+
+function getBibleBookVerseBlocks(bibleBook, verses) {
+  var lastVerseNr = 0;
+  var allBlocks = [];
+  var currentBlock = [];
+
+  // Transform the list of verses into a list of verse blocks (verses that belong together)
+  for (var j = 0; j < verses.length; j++) {
+    var currentVerse = verses[j];
+
+    if (currentVerse.bibleBookShortTitle == bibleBook.shortTitle) {
+
+      if (currentVerse.absoluteVerseNr > (lastVerseNr + 1)) {
+        if (currentBlock.length > 0) {
+          allBlocks.push(currentBlock);
+        }
+        currentBlock = [];
+      }
+      
+      currentBlock.push(currentVerse);
+      lastVerseNr = currentVerse.absoluteVerseNr;
+    }
+  }
+
+  allBlocks.push(currentBlock);
+
+  return allBlocks;
+}
+
+async function renderVerseBlocks(paragraph, bibleBook, verseBlocks) {
+  const bibleTranslationId = app_controller.tab_controller.getTab().getBibleTranslationId();
+  const separator = await i18nHelper.getReferenceSeparator(bibleTranslationId);
+
+  for (let j = 0; j < verseBlocks.length; j++) {
+    const currentBlock = verseBlocks[j];
+
+    const firstVerse = currentBlock[0];
+    const lastVerse = currentBlock[currentBlock.length - 1];
+    
+    // Output the verse reference of this block
+    const bookTitle = await i18nHelper.getSwordTranslation(bibleBook.longTitle);
+    paragraph.addText(bookTitle);
+    paragraph.addText(" " + firstVerse.chapter + separator + firstVerse.verseNr);
+
+    if (currentBlock.length >= 2) { // At least 2 verses, a bigger block
+      let secondRef = "";
+
+      if (lastVerse.chapter == firstVerse.chapter) {
+        secondRef = "-" + lastVerse.verseNr;
+      } else {
+        secondRef = " - " + lastVerse.chapter + separator + lastVerse.verseNr;
+      }
+
+      paragraph.addText(secondRef);
+    }
+    paragraph.addLineBreak();
+
+    for (let k = 0; k < currentBlock.length; k++) {
+      const currentVerse = currentBlock[k];
+      let currentVerseContent = "";
+      const currentVerseNodes = $(jQuery.parseHTML(currentVerse.content));
+      
+      for (let i = 0; i < currentVerseNodes.length; i++) {
+        const currentNode = $(currentVerseNodes[i]);
+        const currentNodeName = currentNode[0].nodeName;
+        // We export everything that is not a DIV
+        // DIV elements contain markup that should not be in the word document
+        if (currentNodeName != 'DIV') {
+          currentVerseContent += currentNode.text();
+        }
+      }
+    
+      paragraph.addText(currentVerse.verseNr + "", { superscript: true });
+      paragraph.addText(" " + currentVerseContent);
+      paragraph.addLineBreak();
+    }
+
+    // Line break after block end
+    paragraph.addLineBreak();
+  }
 }
